+++ windirstat/Options.h
@@ -207,6 +207,7 @@ public:
     static Setting<int> ConfigPage;
     static Setting<int> LanguageId;
     static Setting<int> FileHashAlgorithm;
+    static Setting<int> ProcessPriority;  // 0=Low, 1=Normal, 2=High
     static Setting<int> LargeFileCount;
     static Setting<int> MinimizeViewThreshold;
     static Setting<int> ScanningThreads;